Added support for unqualified form for elements
"unqualified" form means that elements from the target namespace of a schema are
not required to be qualified with the namespace prefix. "unqualified" is the
default form. It can be changed by setting a `form="qualified"` attribute on the
element or a `elementFormDefault="qualified"` attribute on the schema, in which
case it'll apply to all elements in the schema.

Before this change lolsoap always used qualified form regardless of the
attributes in the schema and so all elements in requests were always prefixed
with the namespace.

After this change only elements for which qualified form is specified will be
prefixed.