Use frozen_string_literal
When processing large WSDL's, this can provide non-trival boosts in
speed/memory usage for no cost.

In a real-life scenario where the WSDL is defined by a 15,000 line XML
document, the `memory_monitor` gem showed that the process as a whole
allocated the following amounts of memory:

Before: 54,613,018 bytes
After:  51,802,458 bytes

This is a 2.8MB, or 5% saving.

As a point of interest, the largest saving came from ':' in the
`namespace_and_name` method.